New FileLocation class to handle filename/URL parsing and formatting
Introduces a new class FileLocation, which describes the location of
dive data. The goal of the class is to provide a single place where
the parsing/formating of filenames and URLs is performed. Moreover,
it explicitly stores information on the type instead of encoding it
in the name string. A further goal is to remove the number of global
variables and to detangle the interplay of the git storage code with
the preferences struct.

Currently, five types are supported:
 - None (at program startup / after file close)
 - Local file (.xml file in a local filesystem)
 - Cloud (cloud storage git)
 - Cloud offline (offline copy of the cloud storage)
 - Git (general git, local or remote)

In this commit, the existing_file string and the recent files list
of the main window are converted. The parse_file() and save_dives()
functions are split in two versions, git and local file, since the
frontend now knows which one to choose without parsing the filename.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>