Introduce git_state structure
Introduces a structure, git_state, which is passed around in the git
load/store functions. It contains information on the location of the
git archive as well as temporary data, at the moment the number of
failed authorisation attempts.

The structure is C-style, which means that its resources have to be
free()d using a function (free_git_state).

By introducing this state structure, the global variable auth_attempt
and the encoded_email preferences field could be removed.
The latter simplifies the code quite a bit.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>