Return online state from git machinery in git_state struct
If this is a cloud connection and we can't reach the cloud server,
return this information to the caller by setting the is_remote flag
to false. Also, check the cloud server only for actual cloud connections,
not for any https:// git repositories.

Owing to these changes, the preferences flag git_local_only could be
made local to desktop-widgets/mainwindow.cpp and mobile-widgets/qmlmanager.cpp.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>