Return current SHA in git_state
When manipulating git repositories, return the SHA of the latest commit
in the git_state struct. The caller then remembers the SHA in the FileLocation
struct, if needed.

Thus, the global variable saved_git_id could be removed. By overwriting
currentFile, the old SHA is automatically erased, making the explicit
clearing unnecessary in a number of places.

The desktop verion of the load-from-cloud code was changed to use check_git_sha()
in analogy to the mobile code. Thus the corresponding code in parse_file()
could be removed.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>