Remove cloud_git_url from preferences struct
cloud_git_url was a convenience string generated from cloud_base_url.
Since it was used only at a single place [getCloudLocation()], remove
it from the preferences machinery.

Moreover, make the handling of cloud_base_url consistent with the other
strings in the preferences struct. I.e. always keep a copy of the string
and don't point to the data segment.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>