Support dctype/divemode on UDDF import
This will grab the divemode information from the first waypoint within
the dive that includes this info and uses it for dctype.

Note that our dive type is not as fine grained as the one in UDDF. The
latter allows one to switch divemode during the dive when we do not.
However, one could argue that i.e. a rebreather dive should be a
rebreather dive even if one switches to bailout gas at some point...

Fixes #922

Signed-off-by: Miika Turkia <miika.turkia@gmail.com>