mapwidget: revert QtQuick version to 2.0
In commit f3d978b8a5fb6 the QtQuick version was upgraded
to 2.6 (from 2.0 for the mapwidget). This is, apparently,
too agressive, as there are still mainstream distributions
that are still on Qt 5.5.1 for which 2.6 of QtQuick is too
new. And as a sidenote: Qt 5.5.1 was released in October
2015.

So partially revert commit f3d978b8a5fb6.

Fixes: #978

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>