Add pdf version of user manual
A friend and Subsurface user asked for a PDF version of the
user manual with a working table of contents (that would
not be there when printing out the html).

It took me (on my Mac) to install some asciidoc extension to
generate this, so here it is.

I think it would be nice to offer this to our users. But I am not sure
what is the best way to do this. So I commit this 2.6MB binary
file but feel free to argue that it should be compiled form
the user-manual.txt locally. But then everybody would have to
install such an extension.

Signed-off-by: Robert C. Helling <helling@atdotde.de>