mobile: add localOnly parameter to QMLManager::openLocalThenRemote()
Instead of using the global preferences variable git_local_only to
decide whether to open the remote repository, let the caller decide.

For loading, git_local_only is now set immediately before the call
to the IO code.

In particular this means that the variable currentGitLocalOnly can
be removed, because the old sync-to-cloud status is not overwritten.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>