Mobile: build system, new Kirigami SHA and ...
... fix of a resulting issue. While most new commits are QML
improvements, there is a change in the Kirigami build related
to static building (like we do), and Cmake restyles to make
things more Qt compliant.

As now, the (generated) qrc_kirigami.cpp is included from
kirigamiplugin.cpp, for static builds, our build failed on
double defined symbols. For details of the relevant Kirigami
commmit see 8c34dde11.

The fix on our side is simple. Do not add the resource at our
end, as it is already on the Kirigami side.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>