Mobile: fix build after new Kirigami sha
This is definitely a hack.

Do not include the Kirigami resources (on static build). It causes
double defined symbols in our setting. I would like a nicer fix for this
issue, but failed to find one. For example, not adding the resource in
our build causes the qrc file not to be generated. Manual generation
of the resource file (using rcc) introduces the double symbols again.
so it seems some Kirigami weirdness (but their staticcmake example compiles
correctly).

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>